+++ tests/testthat/test_printf.R
@@ -21,12 +21,14 @@ test_that("warningf", {
   f = function() warningf("123")
   # "Warning: " not caught by gives_warning
   expect_warning(f(), "123")
+  expect_warning(warningf("abc"), "abc")
 })
 
 test_that("stopf", {
   expect_error(stopf("xxx%ixxx", 123), "xxx123xxx")
   f = function() stopf("123")
   expect_error(f(), "123")
+  expect_error(stopf("abc"), "abc")
 })
 
 test_that("non-leanified call is printed", {